Baccarat seems to work. Double payout ratio

+++ Goofbot/UtilClasses/Bets/BaccaratBet.cs
@@ -1,6 +1,6 @@
 ﻿namespace Goofbot.UtilClasses.Bets;
 
-internal class BaccaratBet(long typeID, long payoutRatio, string betName)
+internal class BaccaratBet(long typeID, double payoutRatio, string betName)
     : Bet(typeID, payoutRatio, betName)
 {
 }
