+++ Goofbot/UtilClasses/Bets/RouletteBet.cs
@@ -1,6 +1,6 @@
 ﻿namespace Goofbot.UtilClasses.Bets;
 
-internal class RouletteBet(long typeID, long payoutRatio, string betName)
+internal class RouletteBet(long typeID, double payoutRatio, string betName)
     : Bet(typeID, payoutRatio, betName)
 {
 }
